+++ apps/scouts/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.90.13
+
+### Patch Changes
+
+-   Updated dependencies [[`7e90089f517908562becf72eb3831e9208232278`](https://github.com/learningeconomy/LearnCard/commit/7e90089f517908562becf72eb3831e9208232278)]:
+    -   @learncard/helpers@1.3.2
+    -   @learncard/lca-api-plugin@1.2.11
+    -   @learncard/react@2.9.28
+    -   @learncard/sss-key-manager@0.1.10
+
 ## 1.90.12
 
 ### Patch Changes